Create a setup tasks to try out the sample
One of the requirements to try out the sample is to have ownership of
the package name in the account used to login in AS.

The plugin checks if the account has ownership of the package name
defined in the AndroidManifest.xml. To change this one some extra steps
are required (i.e: R files, applicationId...)

The tasks takes the package name as argument and replaces it in the
proper places.

i.e: ./gradlew setupPackageName -PpackageName="com.package.sample"

It also offer a parameter "disableFirebase" to remove the google service
plugin and comment the Firebase code so the user does not have to
add the google-service.json

Change-Id: Ibb4d32146499e6febb45e6e0254d4162e07a9ca4