Remove setup.gradle task
App Actions Plugin uses now the build.gradle applicationId property,
meaning that is not necessary anymore to change the packageName.

In order to setup the sample only applicationId needs to be change, with
the option to disable Firebase by disabling the flag in app/build.gradle

README updated with new instructions.

Change-Id: I4ef21cba79b11d0f48b872c280711452895ccb93