Handle deeplink when app is on Foreground
The activity launchMode is set to singleTask, avoiding multiple
instances of the FitMainActivity to be created.

Instead onNewIntent(intent) is called with the received intent.

Move the handleIntent code into a common method that will be called from
onCreate and onNewIntent

https: //developer.android.com/guide/topics/manifest/activity-element#lmode
Change-Id: Ibac769238d9bdd8d620703ae9e2a349e9faa1252