Avoid using applicationId for the SliceProvider authority
Since users of the sample will have to change the applicationId to
test the app actions with their own id, the slices won't work unless
they change the authority in the actions.xml to match the applicationId

To avoid adding an extra step, remove the usage of the application id
placeholder in the provider authority and hardcode it.

This means that you can't have multiple instance of this app with
different applicationId installed, since it can only be one provider
with the same authority.

Change-Id: Iceb5e5c9941394a3b5e4e68ad3546bf95b53e2d1